change dmultithread fixedbugs

+++ getSignals.py
@@ -16,6 +16,10 @@ import itertools
 from functools import partial
 from datetime import datetime
 import indicators as ind
+import os.path
+from os import path
+
+
 
 candles = pd.read_csv("XBTUSD-1m-data.csv", sep=',')
 t_symbol = None
@@ -24,31 +28,42 @@ t_candleamount = None
 
 def saveATR(candleamount, params=[], fillna=True, symbol='XBTUSD'):
     for i in params:
+        if path.exists('IndicatorData//' + symbol + '//ATR//' + "p" + str(i) + ".csv"):
+              continue      
         df = ind.atrseries(candles, candleamount, i)
         print(df)
         df.to_csv('IndicatorData//' + symbol + '//ATR//' + "p" + str(i) + ".csv", mode='w')
 
 def saveKeltnerBands(candleamount, params=[], symbol='XBTUSD'):
+    
+
     for i in params:
+        if path.exists('IndicatorData//' + symbol + '//Keltner//' + "BANDS_kp" + str(i[0]) + "_sma" + str(i[1]) + '.csv'):
+            continue
         df = ind.get_keltner_bands(candles, candleamount=candleamount, kperiod=i[0], ksma=i[1])
         print(df)
         df.to_csv('IndicatorData//' + symbol + '//Keltner//' + "BANDS_kp" + str(i[0]) + "_sma" + str(i[1]) + '.csv', mode='w')
 
 def saveKeltnerSignals(candleamount, params=[], symbol='XBTUSD'):
+
     for i in params:
+        if path.exists('IndicatorData//' + symbol + '//Keltner//' + "SIGNALS_kp" + str(i[0]) + "_sma" + str(i[1]) + '.csv'):
+            continue
         signals =  ind.get_keltner_signals(candles, candleamount=candleamount, kperiod=i[0], ksma=i[1])
         df = pd.Series(signals)
         print(df)
         df.to_csv('IndicatorData//' + symbol + '//Keltner//' + "SIGNALS_kp" + str(i[0]) + "_sma" + str(i[1]) + '.csv', mode='w', index=False)
 
 def saveEngulf_thread(params):
+    if path.exists('IndicatorData//' + t_symbol + '//Engulfing//' + "SIGNALS_t" + str(params[0]) + "_ignoredoji" + str(params[1]) + '.csv'):
+        return
+              
     signals = ind.get_engulf_signals(t_e_candles, t_candleamount, params)
     df = pd.Series(signals)
     df.to_csv('IndicatorData//' + t_symbol + '//Engulfing//' + "SIGNALS_t" + str(params[0]) + "_ignoredoji" + str(params[1]) + '.csv', mode='w', index=False)
     return("thread-done")
 
 def saveEngulfingSignals(candleamount, params=[], symbol='XBTUSD'):
-    start = time.time()
     global t_e_candles
     global t_symbol
     global t_candleamount
@@ -61,10 +76,9 @@ def saveEngulfingSignals(candleamount, params=[], symbol='XBTUSD'):
     #DO NOT REMOVE THIS PRINT, IT IS NEEDED TO FINISH THE MULTITHREAD
     result = list(results)
     print(result)
-    end = time.time()
-    print(end - start)
+
     return(result)
-    return(result)
+
 #Examples
 #saveKeltnerBands(100, [10,1], [True, False])
 #saveATR(100, [1,20,30])
